rename package name from mylibrary into viewModel
